requirements.txt
# All dependencies (requires Python 3.10+)
PyGithub>=2.1.1
jira>=3.5.0
openpyxl>=3.1.2
openai>=1.39.0
click>=8.1.7
pyyaml>=6.0.1
python-dateutil>=2.8.2
python-dotenv>=1.0.0
requests>=2.31.0

